Revert "renamed selector.ts -> selectors.ts"
This reverts commit aaf8650a3745443e55ef12ecd07a952a3e41a7e9.
