+++ src/store/slices/offers/selector.ts
